# api.yml
openapi: '3.0.2'
info:
  title: Melanoma API
  version: '1.0'
servers:
  - url: http://localhost:3000 # TODO: update to actual url
paths:
...
  /user/{idUser}/reminder/{idLesion}:
    parameters:
      - $ref: '#/components/parameters/idUser'
      - $ref: '#/components/parameters/idLesion'
    post:
      description: Adds a new reminder to the specified user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reminder'
      responses:
        '200':
          $ref: "#/components/responses/OK"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '400':
          $ref: "#/components/responses/BadRequest"
  /user/{idUser}/reminder/{idReminder}:
    parameters:
      - $ref: '#/components/parameters/idUser'
      - $ref: '#/components/parameters/idReminder'
    get:
      responses:
        '200':
          description: The reminder is returnd successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/reminder'
        '404':
          $ref: "#/components/responses/NotFound"
        '401':
          $ref: "#/components/responses/Unauthorized"
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/reminder'
      responses:
        '200':
          $ref: "#/components/responses/OK"
        '404':
          $ref: "#/components/responses/NotFound"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '400':
          $ref: "#/components/responses/BadRequest"
    delete:
      responses:
        '200':
          $ref: "#/components/responses/OK"
        '401':
          $ref: "#/components/responses/Unauthorized"
  /user/{idUser}/reminder/discard/{idReminder}:
    parameters:
      - $ref: '#/components/parameters/idUser'
      - $ref: '#/components/parameters/idReminder'
    post:
      description: Discards a reminder and creates a new one based on the cycleLength
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/discard'
      responses:
        '200':
          $ref: "#/components/responses/OK"
        '404':
          $ref: "#/components/responses/NotFound"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '400':
          $ref: "#/components/responses/BadRequest"
...
components:
  parameters:
    idLesion:
      name: idLesion
      in: path
      required: true
      schema:
        type: integer
...
    idUser:
      name: idUser
      in: path
      required: true
      schema:
        type: integer
...
    idReminder:
      name: idReminder
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    'OK':
      description: 'Succesful request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/apiResult'
    'Unauthorized':
      description: 'User is not authorized to perform the action'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/apiResult'
    'BadRequest':
      description: 'Request is not valid'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/apiResult'
    'NotFound':
      description: 'The requested resource does not exist'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/apiResult'
  schemas:
    apiResult:
      type: object
      properties:
        result:
          type: boolean
        message:
          type: string
...
    reminder:
      type: object
      properties:
        id:
          type: number
        targetTimeStamp:
          type: string
        lesionID:
          type: number
        cycleLength:
          type: number
...
    discard:
      type: object
      properties:
        erase:
          type: bool